@@ -6,15 +6,17 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1  #应用或包的版本号。
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 #应用或包依赖的其它包或插件。
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  english_words: ^3.1.3
-  fluttertoast: ^3.0.3
+  cupertino_icons: ^1.0.4
+  english_words: ^4.0.0
+  fluttertoast: ^8.0.8
+  google_maps_flutter: ^2.1.1
 
 #开发环境依赖的工具包（而不是flutter应用本身依赖的包）。
 dev_dependencies:
