@@ -6,13 +6,13 @@ class ShareDataWidget extends InheritedWidget{
 
   ShareDataWidget({
    @required this.data,
-    Widget child
-  }) : super(child : child);
+    Widget? child
+  }) : super(child : child!);
 
-  final int data; //需要在子树中共享的数据，保存点击次数
+  final int? data; //需要在子树中共享的数据，保存点击次数
 
   //定义一个便捷方法，方便子树中的widget获取共享数据
-  static ShareDataWidget of(BuildContext context) {
+  static ShareDataWidget? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<ShareDataWidget>();
   }
 
