@@ -57,7 +57,7 @@ class ScrollControllerTestRouteState extends State<ScrollControllerTestRoute> {
               _progress = "${(progress * 100).toInt()}%";
             });
             print("BottomEdge: ${notification.metrics.extentAfter == 0}");
-            //return true; //放开此行注释后，进度条将失效
+            return false; //放开此行注释后，进度条将失效
           },
 
           child: Stack(
