@@ -102,7 +102,7 @@ class _FormTestRouteState  extends State<FormFieldTestRoute>{
                 ),
                 //校验用户名
                 validator: (v){
-                  return v.trim().length>0 ? null : "用户名或邮箱不能为空";
+                  return v!.trim().length>0 ? null : "用户名或邮箱不能为空";
                 },
               ),
 
@@ -120,7 +120,7 @@ class _FormTestRouteState  extends State<FormFieldTestRoute>{
                 obscureText: true,
                 //校验密码
                 validator: (v){
-                  return v.trim().length>5 ? null : "密码不能少于6位";
+                  return v!.trim().length>5 ? null : "密码不能少于6位";
                 },
               ),
 
